prettify function moved to utils.py
